+++ SocialApp/Server/Interfaces/ICookingPageController.cs
@@ -5,8 +5,8 @@ namespace Server.Interfaces
 {
     public interface ICookingPageController
     {
-        ActionResult<CookingPage> GetByUserId(int userId);
-        IActionResult UpdateUserCookingSkill(int userId, int cookingSkillId);
+        ActionResult<CookingPage> GetByUserId(long userId);
+        IActionResult UpdateUserCookingSkill(long userId, int cookingSkillId);
         ActionResult<CookingSkill> GetCookingSkillByDescription(string description);
     }
 }
